+++ Coimbra/ManagedPool.cs
@@ -1,8 +1,6 @@
 ﻿using JetBrains.Annotations;
 using System;
-using System.Collections;
 using System.Collections.Generic;
-using System.Runtime.CompilerServices;
 using UnityEngine;
 using UnityEngine.Scripting;
 using Object = UnityEngine.Object;
@@ -219,13 +217,14 @@ namespace Coimbra
     }
 
     /// <summary>
-    /// Static implementation of <see cref="ManagedPool{T}"/> for objects with a default constructor. It also has special treatment for <see cref="IDictionary"/>, <see cref="IList"/>, <see cref="IDisposable"/>, and <see cref="Object"/>.
+    /// Static implementation of <see cref="ManagedPool{T}"/> for objects with a default constructor. It also has special treatment <see cref="IDisposable"/> and <see cref="Object"/> types to correctly delete those.
     /// </summary>
     [Preserve]
-    public static class ManagedPool
+    [SharedManagedPool("Value", "Instance")]
+    public static partial class ManagedPool
     {
         [Preserve]
-        internal static class Instance<T>
+        private static class Instance<T>
             where T : class, new()
         {
             internal static readonly ManagedPool<T> Value;
@@ -259,71 +258,7 @@ namespace Coimbra
                 }
 
                 Value = new ManagedPool<T>(createCallback, disposeCallback);
-
-                if (typeof(IDictionary).IsAssignableFrom(typeof(T)))
-                {
-                    Value.OnPush += delegate(T obj)
-                    {
-                        ((IDictionary)obj).Clear();
-                    };
-                }
-
-                if (typeof(IList).IsAssignableFrom(typeof(T)))
-                {
-                    Value.OnPush += delegate(T obj)
-                    {
-                        ((IList)obj).Clear();
-                    };
-                }
             }
-        }
-
-        /// <inheritdoc cref="ManagedPool{T}.MaxCapacity"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int GetMaxCapacity<T>()
-            where T : class, new()
-        {
-            return Instance<T>.Value.MaxCapacity;
-        }
-
-        /// <inheritdoc cref="ManagedPool{T}.PreloadCount"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static int GetPreloadCount<T>()
-            where T : class, new()
-        {
-            return Instance<T>.Value.PreloadCount;
-        }
-
-        /// <inheritdoc cref="ManagedPool{T}.Initialize"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void Initialize<T>(int? preloadCount = null, int? maxCapacity = null)
-            where T : class, new()
-        {
-            Instance<T>.Value.Initialize(preloadCount, maxCapacity);
-        }
-
-        /// <inheritdoc cref="ManagedPool{T}.Pop()"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static T Pop<T>()
-            where T : class, new()
-        {
-            return Instance<T>.Value.Pop();
-        }
-
-        /// <inheritdoc cref="ManagedPool{T}.Pop()"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static ManagedPool<T>.Instance Pop<T>([NotNull] out T instance)
-            where T : class, new()
-        {
-            return Instance<T>.Value.Pop(out instance);
-        }
-
-        /// <inheritdoc cref="ManagedPool{T}.Push"/>
-        [MethodImpl(MethodImplOptions.AggressiveInlining)]
-        public static void Push<T>([NotNull] in T instance)
-            where T : class, new()
-        {
-            Instance<T>.Value.Push(in instance);
         }
     }
 }