+++ Roslyn~/Coimbra.Roslyn/CoimbraTypes.cs
@@ -1,0 +1,11 @@
+﻿namespace Coimbra.Roslyn
+{
+    public static class CoimbraTypes
+    {
+        public const string ManagedPoolClass = "ManagedPool";
+
+        public const string Namespace = "Coimbra";
+
+        public const string SharedManagedPoolAttribute = "SharedManagedPoolAttribute";
+    }
+}