+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [10.2.0] - 2022-05-14
+
+- Added easy way to add new shared `ManagedPool` instances through `SharedManagedPoolAttribute`.
+- Added some specific pool collections: `DictionaryPool`, `HashSetPool`, `ListPool`, `QueuePool`, `StackPool`, `StringBuilderPool`.
+- Removed support for `IList` and `IDictionary` in `ManagedPool` class, use the respective specific pool instead (see `COIMBRA0010`).
+
 ## [10.1.2] - 2022-05-13
 
 - Added `Object.Destroy` extension method for safely destroy object that could be a `GameObject` or an `Actor`.